+++ app/controllers/HolidayController.scala
@@ -19,20 +19,12 @@ class HolidayController extends Controller {
     Ok(main(testForm))
   }
 
-  val test = HolidayData
-
   def submit = Action { implicit request =>
     testForm.bindFromRequest().fold(
       formWithErrors => BadRequest(main(testForm)), // should be errors here.
-      name => Ok(hellothere(holidayRemaining(name)))
+      name => Ok(hellothere(HolidayData.holidayRemaining(name)))
     )
   }
-
-  def holidayRemaining(name: String): Int = {
-    val personInfo: List[(String, Int)] = HolidayData.remainingHolidays.filter(_._1 == name)
-    personInfo.head._2
-  }
-
 }
 
 object HolidayController extends HolidayController
